PUBDEV-7622 CoxPH concordance

+++ h2o-genmodel/src/main/java/hex/genmodel/attributes/metrics/MojoModelMetricsRegressionCoxPH.java
@@ -4,5 +4,5 @@ import hex.genmodel.attributes.SerializedName;
 
 public class MojoModelMetricsRegressionCoxPH extends MojoModelMetricsRegression {
   @SerializedName("concordance")
-  public long concordance;  
+  public double concordance;  
 }
